requirements.txt
pip install https://www.djangoproject.com/download/1.7.b4/tarball/
MySQL-python==1.2.5
django-rest-swagger==0.1.14
djangorestframework==2.3.14
wsgiref==0.1.2
